Move xdg popups to xdg header and create them for toplevel views
