--- src/desktop/xdg_popup.c
@@ -1,78 +1,0 @@
-#include "desktop/xdg_popup.h"
-
-#include <stdlib.h>
-#include <assert.h>
-
-#include <wayland-server-core.h>
-#include <wayland-util.h>
-
-#include <wlr/types/wlr_xdg_shell.h>
-
-#include "desktop/tree/node.h"
-
-#include "util/log.h"
-#include "util/wl_macros.h"
-
-//another xdg_popup? oh woaw
-static void e_xdg_popup_new_popup(struct wl_listener* listener, void* data)
-{
-    struct e_xdg_popup* popup = wl_container_of(listener, popup, new_popup);
-    struct wlr_xdg_popup* new_xdg_popup = data;
-
-    e_log_info("popup creates new popup!");
-    e_xdg_popup_create(new_xdg_popup, popup->scene_tree);
-}
-
-//new surface state got committed
-static void e_xdg_popup_commit(struct wl_listener* listener, void* data)
-{
-    struct e_xdg_popup* popup = wl_container_of(listener, popup, commit);
-    
-    if (popup->xdg_popup->base->initial_commit)
-    {
-        //TODO: ensure the popup window isn't positioned offscreen
-        wlr_xdg_surface_schedule_configure(popup->xdg_popup->base);
-    }
-}
-
-//xdg_popup got destroyed
-static void e_xdg_popup_destroy(struct wl_listener* listener, void* data)
-{
-    struct e_xdg_popup* popup = wl_container_of(listener, popup, destroy);
-
-    SIGNAL_DISCONNECT(popup->new_popup);
-    SIGNAL_DISCONNECT(popup->commit);
-    SIGNAL_DISCONNECT(popup->destroy);
-
-    free(popup);
-}
-
-struct e_xdg_popup* e_xdg_popup_create(struct wlr_xdg_popup* xdg_popup, struct wlr_scene_tree* parent)
-{
-    assert(xdg_popup && parent);
-
-    struct e_xdg_popup* popup = calloc(1, sizeof(*popup));
-
-    if (popup == NULL)
-    {
-        e_log_error("failed to allocate xdg_popup");
-        return NULL;
-    }
-
-    popup->xdg_popup = xdg_popup;
-
-    //create popup window's scene tree, and add popup to scene tree of parent
-    popup->scene_tree = wlr_scene_xdg_surface_create(parent, xdg_popup->base);
-    e_node_desc_create(&popup->scene_tree->node, E_NODE_DESC_XDG_POPUP, popup);
-
-    //allows further popup window scene trees to add themselves to this popup window's scene tree
-    xdg_popup->base->data = popup->scene_tree;
-
-    //events
-
-    SIGNAL_CONNECT(xdg_popup->base->events.new_popup, popup->new_popup, e_xdg_popup_new_popup);
-    SIGNAL_CONNECT(xdg_popup->base->surface->events.commit, popup->commit, e_xdg_popup_commit);
-    SIGNAL_CONNECT(xdg_popup->events.destroy, popup->destroy, e_xdg_popup_destroy);
-
-    return popup;
-}
